feat : S3, Redis 분리
