+++ p-application/src/main/java/kr/co/popoolserver/consumer/products/PaymentController.java
@@ -1,4 +1,4 @@
-package kr.co.popoolserver.controllers.products;
+package kr.co.popoolserver.consumer.products;
 
 import io.swagger.annotations.ApiOperation;
 import kr.co.popoolserver.error.model.ResponseFormat;
