+++ p-domain-redis/src/main/java/kr/co/popoolserver/DomainRedisApplication.java
@@ -1,0 +1,11 @@
+package kr.co.popoolserver;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class DomainRedisApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(DomainRedisApplication.class, args);
+    }
+}
