+++ p-domain-redis/src/main/java/kr/co/popoolserver/service/RedisService.java
@@ -1,9 +1,10 @@
-package kr.co.popoolserver.infrastructure.redis;
+package kr.co.popoolserver.service;
 
 import kr.co.popoolserver.error.exception.JwtTokenExpiredException;
 import kr.co.popoolserver.error.model.ErrorCode;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.data.redis.core.ValueOperations;
 import org.springframework.stereotype.Service;
@@ -12,27 +13,23 @@ import java.util.concurrent.TimeUnit;
 
 @Service
 @RequiredArgsConstructor
-public class RedisServiceImpl implements RedisService {
+public class RedisService {
 
     private final RedisTemplate<String, String> redisTemplate;
 
-    @Override
     public String getValue(String key) {
         ValueOperations<String, String> valueOperations = redisTemplate.opsForValue();
         return valueOperations.get(key);
     }
 
-    @Override
     public void createData(String key, String value, long expired) {
         redisTemplate.opsForValue().set(key, value, expired, TimeUnit.MILLISECONDS);
     }
 
-    @Override
     public void deleteData(String key) {
         redisTemplate.delete(key);
     }
 
-    @Override
     public void checkValue(String refreshToken, String redisToken) {
         if(!redisToken.equals(redisToken)) throw new JwtTokenExpiredException(ErrorCode.FAIL_EXPIRE);
     }
