+++ p-domain-redis/src/main/resources/application.yml
@@ -1,0 +1,4 @@
+spring:
+  profiles:
+    include:
+      - local
