+++ p-domain-s3/build.gradle
@@ -1,0 +1,18 @@
+plugins {
+    id 'java'
+}
+
+group 'kr.co.popoolserver'
+
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.7.0'
+    testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.7.0'
+}
+
+test {
+    useJUnitPlatform()
+}
