+++ p-domain-s3/src/main/java/kr/co/popoolserver/DomainS3Application.java
@@ -1,0 +1,11 @@
+package kr.co.popoolserver;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class DomainS3Application {
+    public static void main(String[] args) {
+        SpringApplication.run(DomainS3Application.class, args);
+    }
+}
