+++ p-domain-s3/src/main/java/kr/co/popoolserver/dto/S3Dto.java
@@ -1,0 +1,30 @@
+package kr.co.popoolserver.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import org.springframework.http.HttpHeaders;
+
+public class S3Dto {
+    @Builder
+    @Getter
+    @AllArgsConstructor
+    public static class DOWNLOAD {
+        private byte[] bytes;
+
+        private HttpHeaders httpHeaders;
+    }
+
+    @Builder
+    @Getter
+    @AllArgsConstructor
+    public static class CONVERT{
+        private String fileUrl;
+
+        private String fileName;
+
+        private String fileExtension;
+
+        private long fileSize;
+    }
+}
