+++ settings.gradle
@@ -3,4 +3,7 @@ include 'p-application'
 include 'p-internal'
 include 'p-domain-admin'
 include 'p-domain-consumer'
+include 'p-domain-redis'
+include 'p-domain-s3'
 include 'p-core'
+
